finishing up some refactoring touches
