+++ hacker-stories/src/App.js
@@ -1,51 +1,14 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import Search from './Search'
 import Movies from './Movies'
+import List from './List'
 
 const App = () => {
-  const list = [
-    {
-    title: "Halloween",
-    director: "John Carpenter",
-    releaseYear: 1978,
-    url: "https://en.wikipedia.org/wiki/Halloween_(1978_film)",
-    objectID: 1
-    },
-    {
-      title: "A Nightmare on Elm Street",
-      director: "Wes Craven",
-      releaseYear: 1984,
-      url: "https://en.wikipedia.org/wiki/A_Nightmare_on_Elm_Street",
-      objectID: 2
-    },
-    {
-      title: "Texas Chainsaw Massacre",
-      director: "Tobe Hooper",
-      releaseYear: 1974,
-      url: "https://en.wikipedia.org/wiki/The_Texas_Chain_Saw_Massacre",
-      objectID: 3
-    },
-    {
-      title: "The Thing",
-      director: "John Carpenter",
-      releaseYear: 1982,
-      url: "https://www.imdb.com/title/tt0084787/",
-      objectID: 4
-    },
-    {
-      title: "The Fly",
-      director: "David Cronenberg",
-      releaseYear: 1986,
-      url: "https://www.imdb.com/title/tt0091064/?ref_=nv_sr_srsg_0",
-      objectID: 5
-    }
-  ]
-  
-  const [searchTerm, setSearchTerm] = React.useState(
+  const [searchTerm, setSearchTerm] = useState(
     localStorage.getItem('search') || "Halloween"
     );
 
-  React.useEffect(() => {
+  useEffect(() => {
     localStorage.setItem('search', searchTerm)
   }, [searchTerm])
 
@@ -54,7 +17,7 @@ const App = () => {
     setSearchTerm(event.target.value)
   }
 
-  const filteredMovies = list.filter((movie) => 
+  const filteredMovies = List.filter((movie) => 
     (movie.title.toLowerCase().includes(searchTerm.toLowerCase()))
   );
